make makerpm a little quicker about copying and stuff

+++ makerpm.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/bin/bash -e
 
 ###
 # First stab at a script to build RPMs
@@ -25,44 +25,48 @@ fi
 
 VERSION=1.3.7
 
-if [ -z $JAVA_HOME ]
-then
-echo "*** JAVA_HOME must be set ***"
-exit
+if [ -z $JAVA_HOME ]; then
+	echo "*** JAVA_HOME must be set ***"
+	exit
 fi
+
+export PATH="$JAVA_HOME/bin:$PATH"
+TAR=`which gtar 2>/dev/null || which tar 2>/dev/null`
+RSYNC=`which rsync 2>/dev/null`
 
 echo "==== Building OpenNMS RPMs ===="
 echo
 echo "Version: " $VERSION
 echo "Release: " $RELEASE
-echo "Platform: " $PLATFORM
 echo
 
 echo "=== Build Clean ==="
 
 ./build.sh clean
 
-if [ -d /tmp/opennms-$VERSION-$RELEASE/source ]
-then
-echo "=== Removing existing source directory ==="
-rm -rf /tmp/opennms-$VERSION-$RELEASE/source
+if [ -n "$RSYNC" ]; then
+
+	echo "=== Copying Source to Source Directory ==="
+	$RSYNC -aqr --exclude=.svn --delete ./ /tmp/opennms-$VERSION-$RELEASE/source/
+
+else
+
+	if [ -d /tmp/opennms-$VERSION-$RELEASE/source ]; then
+		echo "=== Removing existing source directory ==="
+		rm -rf /tmp/opennms-$VERSION-$RELEASE/source
+	fi
+
+	echo "=== Create Source Directory ==="
+	mkdir -p /tmp/opennms-$VERSION-$RELEASE/source
+
+	echo "=== Copying Source to Source Directory ==="
+	cp -r * /tmp/opennms-$VERSION-$RELEASE/source/
+
 fi
-
-echo "=== Create Source Directory ==="
-
-mkdir -p /tmp/opennms-$VERSION-$RELEASE/source
-
-echo "=== Copying Source to Source Directory ==="
-
-cp -r * /tmp/opennms-$VERSION-$RELEASE/source/
-
-echo "=== Removing .svn directories from the Source ==="
-
-find /tmp/opennms-$VERSION-$RELEASE/source -type d -name ".svn" -exec rm \-rf {} \;
 
 echo "=== Creating a tar.gz archive of the Source in /usr/src/redhat/SOURCES ==="
 
-tar zcvf /usr/src/redhat/SOURCES/opennms-source-$VERSION-$RELEASE.tar.gz -C /tmp opennms-$VERSION-$RELEASE
+$TAR zcvf /usr/src/redhat/SOURCES/opennms-source-$VERSION-$RELEASE.tar.gz -C /tmp opennms-$VERSION-$RELEASE
 
 echo "=== Building RPMs ==="
